Refactoring frame, field utilities
